requirements.txt
asgiref
Django
djangorestframework
markdown
django-filter
pytz
setuptools
sqlparse
wheel
djangorestframework-simplejwt